+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added in support for import statements.
 * Added in `dependencies` command that prints out a list of the file dependencies.
+* Added in `--libpath` option on the `compile` and `build` commands for referencing extra libraries
+  during compilation
 
 #### v0.2.0.0